DrawModes made, files moved. Bounding box sorta fixed.
