+++ src/DrawModes/CopyMultiMode.ts
@@ -1,15 +1,16 @@
 /**
- * File containing multi node movement event handlers.
+ * File containing copy node movement event handlers.
+ * @author Anusha Tiwari
  * @author Dawn Moore
  */
 
-import {Point} from "./AEG/Point";
-import {AtomNode} from "./AEG/AtomNode";
-import {CutNode} from "./AEG/CutNode";
-import {tree} from "./index";
+import {Point} from "../AEG/Point";
+import {AtomNode} from "../AEG/AtomNode";
+import {CutNode} from "../AEG/CutNode";
+import {tree} from "../index";
 import {offset} from "./DragMode";
 import {drawAtom, redrawTree} from "./DrawUtils";
-import {legalColor, illegalColor} from "./Themes";
+import {legalColor, illegalColor} from "../Themes";
 import {validateChildren, drawAltered, insertChildren, alterAtom} from "./EditModeUtils";
 
 //The initial point the user pressed down.
@@ -23,17 +24,13 @@ let legalNode: boolean;
 
 /**
  * Takes the starting point and sets the lowest node containing that point that is not the sheet to
- * the current node. Removes that node from its parent.
- * @param event The mouse down event while in moveMulti mode.
+ * the current node.
+ * @param event The mouse down event while in copyMulti mode.
  */
-export function moveMultiMouseDown(event: MouseEvent) {
+export function copyMultiMouseDown(event: MouseEvent) {
     startingPoint = new Point(event.x - offset.x, event.y - offset.y);
     currentNode = tree.getLowestNode(startingPoint);
     if (currentNode !== tree.sheet && currentNode !== null) {
-        const currentParent = tree.getLowestParent(startingPoint);
-        if (currentParent !== null) {
-            currentParent.remove(startingPoint);
-        }
         legalNode = true;
     } else {
         legalNode = false;
@@ -44,9 +41,9 @@ export function moveMultiMouseDown(event: MouseEvent) {
  * If the node selected was legal, draws the node with the difference between the starting position
  * and the current position by altering the point of origin. If the node was a cut node also draws
  * all of the children with the same change in location.
- * @param event The mouse move event while in moveMulti mode
+ * @param event The mouse move event while in copyMulti mode
  */
-export function moveMultiMouseMove(event: MouseEvent) {
+export function copyMultiMouseMove(event: MouseEvent) {
     if (legalNode) {
         const moveDifference: Point = new Point(
             event.x - startingPoint.x,
@@ -74,12 +71,12 @@ export function moveMultiMouseMove(event: MouseEvent) {
 
 /**
  * If the current node is a cut node, and all of its children are in a legal position places it
- * in the current position. If it is not in a legal position returns the original node to the tree.
- * If the current node is an atom node and is in a legal position adds it to the tree, otherwise
- * readds the original node in the original place.
- * @param event the mouse up event while in moveMulti mode
+ * in the current position. If it is not in a legal position, copy failed and nothing is inserted.
+ * If the current node is an atom node and is in a legal position inserts it in the tree, otherwise
+ * copy failed and nothing is inserted.
+ * @param event the mouse up event while in copyMulti mode
  */
-export function moveMultiMouseUp(event: MouseEvent) {
+export function copyMultiMouseUp(event: MouseEvent) {
     if (legalNode) {
         const moveDifference: Point = new Point(
             event.x - startingPoint.x,
@@ -89,16 +86,12 @@ export function moveMultiMouseUp(event: MouseEvent) {
         if (currentNode instanceof CutNode) {
             if (validateChildren(currentNode, moveDifference)) {
                 insertChildren(currentNode, moveDifference);
-            } else {
-                tree.insert(currentNode);
             }
         } else if (currentNode instanceof AtomNode) {
             const tempAtom: AtomNode = alterAtom(currentNode, moveDifference);
 
             if (tree.canInsert(tempAtom)) {
                 tree.insert(tempAtom);
-            } else {
-                tree.insert(currentNode);
             }
         }
     }
@@ -107,13 +100,10 @@ export function moveMultiMouseUp(event: MouseEvent) {
 }
 
 /**
- * If the current node is a legal node returns it to the original position.
+ * If the mouse is moved outside of the canvas, sets wasOut to true.
  * Redraws the canvas to clear any drawings not part of the tree.
  */
-export function moveMultiMouseOut() {
-    if (legalNode && currentNode !== null) {
-        tree.insert(currentNode);
-    }
+export function copyMultiMouseOut() {
     legalNode = false;
     redrawTree(tree);
 }
